+++ CHANGELOG.md
@@ -8,7 +8,10 @@ ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: drop support for 32-bit Windows distributables
 * `FEAT`: rename cluster connection terminology to plain connection ([#6158](https://github.com/camunda/camunda-modeler/pull/6158))
+* `FEAT`: report unreachable or temporarily unavailable instance on connection check ([#6188](https://github.com/camunda/camunda-modeler/pull/6188))
 * `FIX`: do not report a connection error while the tab's connection is still being resolved
+* `FIX`: recover connection once an initially unreachable cluster comes back up ([#6188](https://github.com/camunda/camunda-modeler/pull/6188))
+* `FIX`: correct spacing of connection error description in connection chooser ([#6188](https://github.com/camunda/camunda-modeler/pull/6188))
 * `DEPS`: update to `electron@44.0.0`
 
 ## 5.51.1